requirements.txt
anyio==3.4.0
fastapi==0.62.0
fastapi-utils==0.2.1
uvicorn==0.12.3
gunicorn==20.0.4
uvloop==0.14.0
httptools==0.1.1
requests==2.24.0
PyJWT==1.4.2
python-multipart==0.0.5
python-json-logger==0.1.11
aiofiles==0.6.0
redis==3.5.3
minio==7.0.3
xmltodict==0.12.0
numpy==1.20.1
httpx==0.21.1
pydantic==1.8.2
opentelemetry-exporter-jaeger==1.6.2
opentelemetry-instrumentation==0.26b1
opentelemetry-instrumentation-fastapi==0.26b1
opentelemetry-instrumentation-sqlalchemy==0.26b1
opentelemetry-instrumentation-requests==0.26b1
opentelemetry-instrumentation-asyncpg==0.26b1
SQLAlchemy==1.4.27
psycopg2-binary==2.9.2
asyncpg==0.25.0